✨ feat: refactor batch download types and move TrpcSubscriptionCallback to a new file
