+++ src/services/file/index.ts
@@ -1,12 +1,12 @@
 import { Unsubscribable } from '@trpc/server/observable';
 
 import { lambdaClient } from '@/libs/trpc/client';
+import { TrpcSubscriptionCallback } from '@/services/file/type';
 import {
   CheckFileHashResult,
   FileItem,
   QueryFileListParams,
   QueryFileListSchemaType,
-  TrpcSubscriptionCallback,
   UploadFileParams,
 } from '@/types/files';
